refactor(cli): move CLI-specific utilities to CLI directory
- Move profile-display.ts from utils/ to cli/commands/profile/display.ts
- Move first-time.ts from utils/ to cli/first-time.ts
- Keep goodbye-messages.ts in utils/ (used by agents, truly shared)
- Follow established pattern: CLI helpers belong with CLI code
- Update imports in cli/index.ts and profile/index.ts
