+++ src/cli/commands/profile/index.ts
@@ -3,7 +3,7 @@ import chalk from 'chalk';
 import inquirer from 'inquirer';
 import { ConfigLoader } from '../../../utils/config.js';
 import { logger } from '../../../utils/logger.js';
-import { ProfileDisplay } from '../../../utils/profile-display.js';
+import { ProfileDisplay } from './display.js';
 import { ProviderRegistry } from '../../../providers/core/registry.js';
 import { handleAuthValidationFailure } from '../../../providers/core/auth-validation.js';
 import { createLoginCommand, createLogoutCommand, createRefreshCommand } from './auth.js';
